clickhouse: removed ddl; split test functions file; refactored test client
